Split long file in two
